refactor: 서비스 레이어 인터페이스 및 구현체 분리
