+++ payment_service/src/main/java/on/ssgdeal/payment_service/presentation/external/PaymentController.java
@@ -3,9 +3,10 @@ package on.ssgdeal.payment_service.presentation.external;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
+import on.ssgdeal.common.auth.passport.Passport;
 import on.ssgdeal.common.auth.passport.PassportUtil;
 import on.ssgdeal.common.presentation.dto.CommonResponse;
-import on.ssgdeal.payment_service.application.service.ConfirmPaymentService;
+import on.ssgdeal.payment_service.application.service.PaymentProcessorService;
 import on.ssgdeal.payment_service.application.service.dto.response.OrderPaymentResponseDto;
 import on.ssgdeal.payment_service.presentation.external.dto.OrderPaymentRequest;
 import org.springframework.http.ResponseEntity;
@@ -19,7 +20,7 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/api/v1/payment")
 public class PaymentController {
 
-    private final ConfirmPaymentService paymentService;
+    private final PaymentProcessorService paymentProcessorService;
     private final PassportUtil passportUtil;
 
     @PostMapping("/confirm")
@@ -27,11 +28,9 @@ public class PaymentController {
         @RequestBody @Valid OrderPaymentRequest request,
         HttpServletRequest servletRequest
     ) {
-        // TODO: 인증 수정하
-//        final Passport passport = passportUtil.getPassportBy(servletRequest);
-//        final var requestDto = request.toDto(passport.getUserId());
-        final var requestDto = request.toDto(1L);
-        final var responseDto = paymentService.orderPayment(requestDto);
+        final Passport passport = passportUtil.getPassportBy(servletRequest);
+        final var requestDto = request.toDto(passport.getUserId());
+        final var responseDto = paymentProcessorService.orderPayment(requestDto);
         return ResponseEntity.ok(CommonResponse.success(responseDto));
     }
 }
